HAI-2310 Move the remaining classes from application to hakemus
The moved classes are:
- ApplicationData
- ApplicationDeletionResultDto
- ApplicationEntity
- ApplicationRepository

They are moved seperately to make saving git history easier. They will
be renamed in later commits.
