+++ services/hanke-service/src/main/kotlin/fi/hel/haitaton/hanke/hakemus/ApplicationData.kt
@@ -1,4 +1,4 @@
-package fi.hel.haitaton.hanke.application
+package fi.hel.haitaton.hanke.hakemus
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties
 import com.fasterxml.jackson.annotation.JsonSubTypes
@@ -6,15 +6,6 @@ import com.fasterxml.jackson.annotation.JsonTypeInfo
 import com.fasterxml.jackson.annotation.JsonView
 import fi.hel.haitaton.hanke.ChangeLogView
 import fi.hel.haitaton.hanke.NotInChangeLogView
-import fi.hel.haitaton.hanke.hakemus.ApplicationType
-import fi.hel.haitaton.hanke.hakemus.HakemusData
-import fi.hel.haitaton.hanke.hakemus.Hakemusalue
-import fi.hel.haitaton.hanke.hakemus.Hakemusyhteystieto
-import fi.hel.haitaton.hanke.hakemus.JohtoselvitysHakemusalue
-import fi.hel.haitaton.hanke.hakemus.JohtoselvityshakemusData
-import fi.hel.haitaton.hanke.hakemus.KaivuilmoitusAlue
-import fi.hel.haitaton.hanke.hakemus.KaivuilmoitusData
-import fi.hel.haitaton.hanke.hakemus.Laskutusyhteystieto
 import java.time.ZonedDateTime
 
 enum class ApplicationContactType {
