+++ services/hanke-service/src/main/kotlin/fi/hel/haitaton/hanke/hakemus/ApplicationDeletionResultDto.kt
@@ -1,3 +1,3 @@
-package fi.hel.haitaton.hanke.application
+package fi.hel.haitaton.hanke.hakemus
 
 data class ApplicationDeletionResultDto(val hankeDeleted: Boolean)
