+++ services/hanke-service/src/main/kotlin/fi/hel/haitaton/hanke/hakemus/ApplicationEntity.kt
@@ -1,12 +1,7 @@
-package fi.hel.haitaton.hanke.application
+package fi.hel.haitaton.hanke.hakemus
 
 import fi.hel.haitaton.hanke.HankeEntity
 import fi.hel.haitaton.hanke.allu.ApplicationStatus
-import fi.hel.haitaton.hanke.hakemus.ApplicationType
-import fi.hel.haitaton.hanke.hakemus.Hakemus
-import fi.hel.haitaton.hanke.hakemus.HakemusIdentifier
-import fi.hel.haitaton.hanke.hakemus.HakemusMetaData
-import fi.hel.haitaton.hanke.hakemus.HakemusyhteystietoEntity
 import fi.hel.haitaton.hanke.permissions.HankekayttajaEntity
 import io.hypersistence.utils.hibernate.type.json.JsonType
 import jakarta.persistence.CascadeType
