+++ services/hanke-service/src/main/kotlin/fi/hel/haitaton/hanke/hakemus/ApplicationRepository.kt
@@ -1,4 +1,4 @@
-package fi.hel.haitaton.hanke.application
+package fi.hel.haitaton.hanke.hakemus
 
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.Query
